backend/requirements.txt
flask
flask-cors
pandas
openpyxl
ortools
geopy
gunicorn
numpy
sqlalchemy
psycopg2-binary
requests
PyJWT
bcrypt
